requirements.txt
PyYAML
beautifulsoup4
html5lib
